Restructure Microservice
