+++ src/main/java/br/com/hideyoshi/auth/controller/UserController.java
@@ -5,9 +5,9 @@ import br.com.hideyoshi.auth.model.TokenDTO;
 import br.com.hideyoshi.auth.service.AuthService;
 import br.com.hideyoshi.auth.model.UserDTO;
 import br.com.hideyoshi.auth.service.UserService;
-import br.com.hideyoshi.auth.microservice.storageservice.enums.FileTypeEnum;
-import br.com.hideyoshi.auth.microservice.storageservice.model.StorageServiceUploadResponse;
-import br.com.hideyoshi.auth.microservice.storageservice.service.StorageService;
+import br.com.hideyoshi.auth.enums.FileTypeEnum;
+import br.com.hideyoshi.auth.model.microservice.StorageServiceUploadResponse;
+import br.com.hideyoshi.auth.service.microservice.StorageService;
 import br.com.hideyoshi.auth.util.exception.BadRequestException;
 import br.com.hideyoshi.auth.util.guard.UserResourceGuard;
 import br.com.hideyoshi.auth.util.guard.UserResourceGuardEnum;
