+++ src/main/java/br/com/hideyoshi/auth/enums/FileTypeEnum.java
@@ -1,4 +1,4 @@
-package br.com.hideyoshi.auth.microservice.storageservice.enums;
+package br.com.hideyoshi.auth.enums;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import br.com.hideyoshi.auth.util.exception.BadRequestException;
