+++ src/main/java/br/com/hideyoshi/auth/service/AuthService.java
@@ -8,8 +8,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import br.com.hideyoshi.auth.entity.Provider;
 import br.com.hideyoshi.auth.model.UserDTO;
-import br.com.hideyoshi.auth.microservice.storageservice.model.StorageServiceDownloadResponse;
-import br.com.hideyoshi.auth.microservice.storageservice.service.StorageService;
+import br.com.hideyoshi.auth.model.microservice.StorageServiceDownloadResponse;
+import br.com.hideyoshi.auth.service.microservice.StorageService;
 import br.com.hideyoshi.auth.util.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
