+++ src/main/java/br/com/hideyoshi/auth/service/microservice/StorageService.java
@@ -1,11 +1,11 @@
-package br.com.hideyoshi.auth.microservice.storageservice.service;
+package br.com.hideyoshi.auth.service.microservice;
 
+import br.com.hideyoshi.auth.config.StorageServiceConfig;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import br.com.hideyoshi.auth.microservice.storageservice.config.StorageServiceConfig;
-import br.com.hideyoshi.auth.microservice.storageservice.enums.FileTypeEnum;
-import br.com.hideyoshi.auth.microservice.storageservice.model.StorageServiceDownloadResponse;
-import br.com.hideyoshi.auth.microservice.storageservice.model.StorageServiceUploadResponse;
+import br.com.hideyoshi.auth.enums.FileTypeEnum;
+import br.com.hideyoshi.auth.model.microservice.StorageServiceDownloadResponse;
+import br.com.hideyoshi.auth.model.microservice.StorageServiceUploadResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.http.client.methods.HttpDelete;
